requirements.txt
absl-py==2.2.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==10.0.0
babel==2.17.0
boto3==1.38.23
botocore==1.38.23
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cn2an==0.5.22
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==3.24.0
cycler==0.12.1
Cython==3.1.1
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
dtw-python==1.4.4
eng-to-ipa==0.0.2
fastapi==0.115.12
faster-whisper==0.9.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.71.0
grpclib==0.4.7
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langid==1.1.6
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.44.0
loguru==0.7.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.9
melotts==0.1.2
modal==1.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
narwhals==1.41.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.0
openai==1.82.0
openai-whisper==20240930
orjson==3.10.18
packaging==25.0
pandas==1.5.1
pillow==10.4.0
plac==1.4.5
platformdirs==4.3.8
pooch==1.8.2
proces==0.1.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pykakasi==2.2.1
pyparsing==3.2.3
pypinyin==0.50.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.14.0
synchronicity==0.9.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.13.3
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.27.4
txtsplit==1.0.0
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.4.0
uvicorn==0.34.2
wasabi==0.10.1
watchfiles==1.0.5
wavmark==0.0.3
websockets==11.0.3
Werkzeug==3.1.3
whisper-timestamped==1.14.2
wrapt==1.17.2
yarl==1.20.0
